+++ Components/OoliteBase/Source/OOIsNumberLiteral.m
@@ -42,7 +42,7 @@ BOOL OOIsNumberLiteral(NSString *string, BOOL allowSpaces)
 							hadExp = NO,
 							allowDec = YES,
 							hadNumber = NO;
-	unsigned				i, count;
+	NSUInteger				i, count;
 	
 	if (string == nil)  return NO;
