+++ Components/OoliteBase/Source/OORandom.h
@@ -137,7 +137,7 @@ OOINLINE double distanceBetweenPlanetPositions ( int x1, int y1, int x2, int y2)
 {
 	int dx = x1 - x2;
 	int dy = (y1 - y2)/2;
-	int dist = sqrtf(dx*dx + dy*dy);	// N.b. Rounding error due to truncation is desired.
+	int dist = (int)sqrtf(dx*dx + dy*dy);	// N.b. Rounding error due to truncation is desired.
 	return 0.4 * dist;
 }
